implemented last feedbacks
