+++ python-client/giskard/model/utils.py
