+++ python-client/tests/fixtures/enron_multilabel_classification.py
@@ -13,7 +13,7 @@ from sklearn.preprocessing import OneHotEncoder, StandardScaler
 from giskard.core.core import SupportedModelTypes
 from giskard.ml_worker.core.dataset import Dataset
 from giskard.ml_worker.utils.logging import Timer
-from giskard.models.sklearn import SKLearnModel
+from giskard.model.sklearn import SKLearnModel
 from tests import path
 
 logger = logging.getLogger(__name__)
