+++ python-client/tests/fixtures/german_credit_scoring.py
@@ -13,8 +13,8 @@ from sklearn.preprocessing import OneHotEncoder, StandardScaler
 from giskard.core.core import SupportedModelTypes
 from giskard.ml_worker.core.dataset import Dataset
 from giskard.ml_worker.utils.logging import Timer
-from giskard.models.catboost import CatboostModel
-from giskard.models.sklearn import SKLearnModel
+from giskard.model.catboost import CatboostModel
+from giskard.model.sklearn import SKLearnModel
 from tests import path
 
 logger = logging.getLogger(__name__)
