+++ python-client/tests/fixtures/regression.py
@@ -5,7 +5,7 @@ from sklearn.metrics import mean_squared_error
 from giskard.core.core import SupportedModelTypes
 from giskard.ml_worker.core.dataset import Dataset
 from giskard.ml_worker.utils.logging import Timer
-from giskard.models.sklearn import SKLearnModel
+from giskard.model.sklearn import SKLearnModel
 
 
 @pytest.fixture()
